requirements.txt
asgiref==3.4.1
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
Deprecated==1.2.13
Django==3.2.7
django-cors-headers==3.8.0
django-phonenumber-field==5.2.0
django-rest==0.8.7
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jwt==1.2.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
phonenumbers==8.12.31
platformdirs==2.3.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
regex==2021.8.28
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
vonage==2.5.5
wrapt==1.12.1